+++ Tests/KeyboardKitTests/Extensions/UITextDocumentProxy+ContentTests.swift
@@ -20,10 +20,15 @@ class UITextDocumentProxy_ContentTests: QuickSpec {
             proxy = MockTextDocumentProxy()
         }
         
+        func prepareProxy(with preCursorPart: String, _ postCursorPart: String? = nil) {
+            proxy.documentContextBeforeInput = preCursorPart
+            proxy.documentContextAfterInput = postCursorPart
+        }
+        
         describe("is cursor at the beginning of a new sentence") {
             
             func result(for preCursorPart: String) -> Bool {
-                proxy.documentContextBeforeInput = preCursorPart
+                prepareProxy(with: preCursorPart)
                 return proxy.isCursorAtNewSentence
             }
             
@@ -46,7 +51,7 @@ class UITextDocumentProxy_ContentTests: QuickSpec {
         describe("is cursor at the beginning of a new word") {
             
             func result(for preCursorPart: String) -> Bool {
-                proxy.documentContextBeforeInput = preCursorPart
+                prepareProxy(with: preCursorPart)
                 return proxy.isCursorAtNewWord
             }
             
@@ -66,11 +71,65 @@ class UITextDocumentProxy_ContentTests: QuickSpec {
             }
         }
         
+        describe("sentence before input") {
+            
+            func result(for preCursorPart: String, _ postCursorPart: String) -> String? {
+                prepareProxy(with: preCursorPart, postCursorPart)
+                return proxy.sentenceBeforeInput
+            }
+            
+            it("returns last ended sentence, if any") {
+                expect(result(for: "", "")).to(beNil())
+                expect(result(for: "sentence", "")).to(beNil())
+                expect(result(for: "sentence.", "")).to(equal("sentence"))
+                expect(result(for: "sentence!", "")).to(equal("sentence"))
+                expect(result(for: "sentence .", "")).to(equal("sentence "))
+                expect(result(for: "sentence. ", "")).to(beNil())
+                expect(result(for: "sentence. a", "")).to(beNil())
+                expect(result(for: "sentence.a", "")).to(beNil())
+            }
+        }
+        
+        describe("last ended word before input") {
+            
+            func result(for preCursorPart: String, _ postCursorPart: String) -> String? {
+                prepareProxy(with: preCursorPart, postCursorPart)
+                return proxy.wordBeforeInput
+            }
+            
+            it("returns previous word if one exists and the cursor is not at the beginning of a new sentence") {
+                expect(result(for: "word", "")).to(beNil())
+                expect(result(for: "word ", "")).to(equal("word"))
+                expect(result(for: "word. ", "")).to(beNil())
+                expect(result(for: "one two . three ", "")).to(equal("three"))
+                expect(result(for: "word...", "")).to(beNil())
+                expect(result(for: "word,,,", "")).to(beNil())
+                expect(result(for: "word,", "")).to(equal("word"))
+                expect(result(for: "word, ", "")).to(beNil())
+            }
+        }
+        
         describe("sentence delimiter list") {
             
             it("is correctly setup") {
                 let result = proxy.sentenceDelimiters
                 expect(result).to(equal(["!", ".", "?", "\n"]))
+            }
+        }
+        
+        describe("trimmed document context after input") {
+            
+            it("removes whitespace") {
+                proxy.documentContextAfterInput = " foo "
+                expect(proxy.trimmedDocumentContextAfterInput).to(equal("foo"))
+            }
+        }
+        
+        describe("trimmed document context before input") {
+            
+            it("removes whitespace") {
+                proxy.documentContextBeforeInput = " bar "
+                expect(proxy.trimmedDocumentContextBeforeInput).to(equal("bar"))
             }
         }